Fix drop down side menu behavior
